# apps/docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOCA API",
    "version": "1.0.0",
    "description": "The unified public API of the **Museum of Crypto Art** \u2014 collections, artworks (with original-ratio media), 3D exhibition rooms, and the **Art DeCC0s** knowledge base, aggregated behind a single API key.\n\nAll endpoints (except the index) require a MOCA API key sent as `X-API-Key: <key>` or `Authorization: Bearer <key>`.\n\nResponses use the Directus-style envelope: `{ \"data\": \u2026, \"meta\": \u2026 }` on success and `{ \"errors\": [{ \"message\", \"extensions\": { \"code\" } }] }` on failure.",
    "contact": {
      "name": "Museum of Crypto Art",
      "url": "https://museumofcryptoart.com"
    }
  },
  "servers": [
    {
      "url": "https://api.moca.qwellco.de",
      "description": "Production"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/collections": {
      "get": {
        "tags": [
          "Collections"
        ],
        "operationId": "listCollections",
        "summary": "List collections",
        "description": "All published top-level museum collections, each with its child collections.",
        "responses": {
          "200": {
            "description": "Collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Collection"
                      }
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": 1,
                      "name": "Genesis Collection",
                      "title": "The Genesis Collection",
                      "description": "The founding artworks of the museum.",
                      "slug": "genesis",
                      "parent_collection": null,
                      "child_collections": [
                        {
                          "id": 4,
                          "name": "Genesis II",
                          "slug": "genesis-ii"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/collections/{slug}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "operationId": "getCollection",
        "summary": "Get a collection",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "genesis"
          }
        ],
        "responses": {
          "200": {
            "description": "Collection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Collection"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/artworks": {
      "get": {
        "tags": [
          "Artworks"
        ],
        "operationId": "listArtworks",
        "summary": "List / search artworks",
        "description": "Paged artworks across the museum. Media is normalized server-side: dead URLs are revived, OpenSea's square-cropped CDN variants are swapped for the **original files**, and `ratio` only ever derives from trustworthy dimensions \u2014 render `media.url` at `ratio` and you are showing the artwork the way the artist made it.",
        "parameters": [
          {
            "name": "collection",
            "in": "query",
            "description": "Comma-separated collection slug(s) to scope the query.",
            "schema": {
              "type": "string"
            },
            "example": "genesis"
          },
          {
            "name": "search",
            "in": "query",
            "description": "Match against artwork title or artist name.",
            "schema": {
              "type": "string"
            },
            "example": "butterfly"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Artworks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Artwork"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": 3643,
                      "name": "Tree of Origins",
                      "artist_name": "Example Artist",
                      "collection": "genesis",
                      "media": {
                        "url": "https://ipfs.pixura.io/ipfs/QmfHBdQHPYKC3P2CdotUxDZzC1AAUWnLks6U7SBWz1RbGD",
                        "type": "image",
                        "content_type": "image/png",
                        "width": 1026,
                        "height": 1431
                      },
                      "animation": null,
                      "ratio": 0.717,
                      "opensea_url": "https://opensea.io/assets/ethereum/0x41a3\u2026/12"
                    }
                  ],
                  "meta": {
                    "total": 5210,
                    "page": 1,
                    "limit": 25
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/artworks/{id}": {
      "get": {
        "tags": [
          "Artworks"
        ],
        "operationId": "getArtwork",
        "summary": "Get an artwork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 3643
          }
        ],
        "responses": {
          "200": {
            "description": "Artwork",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Artwork"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/rooms": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "operationId": "listRooms",
        "summary": "List 3D rooms",
        "description": "The museum's 3D exhibition architecture. Each room ships a GLB model (`model_url`) whose `Slot_001\u2026Slot_NNN` placeholder meshes mark wall positions for hanging artworks \u2014 the same convention the museum's own world builder uses.",
        "responses": {
          "200": {
            "description": "Rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Room"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/decc0s": {
      "get": {
        "tags": [
          "DeCC0s"
        ],
        "operationId": "listDecc0s",
        "summary": "List / search DeCC0s",
        "description": "The 10,000 Art DeCC0 entities, aggregated live from the DeCC0s knowledge base (api.decc0s.com) with a short server-side cache. List responses return a lightweight field set; use `fields` to choose your own, or fetch `/v1/decc0s/{id}` for the full record.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "search",
            "in": "query",
            "description": "Full-text search across the knowledge base.",
            "schema": {
              "type": "string"
            },
            "example": "Cairo"
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated field selection (the heavyweight `agent_profiles` / `moltbot` blobs are list-excluded regardless).",
            "schema": {
              "type": "string"
            },
            "example": "id,name,description,owner"
          }
        ],
        "responses": {
          "200": {
            "description": "DeCC0s",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Decc0Summary"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Decc0ListMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        }
      }
    },
    "/v1/decc0s/{id}": {
      "get": {
        "tags": [
          "DeCC0s"
        ],
        "operationId": "getDecc0",
        "summary": "Get a DeCC0",
        "description": "One Art DeCC0 entity by token id (1\u201310000). Add `include=profiles` for the full agent persona blobs (`agent_profiles`, `moltbot` \u2014 hundreds of KB), and `include=codex` to embed the character's codex lore document.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000
            },
            "example": 1
          },
          {
            "name": "include",
            "in": "query",
            "description": "Comma-separated extras: `profiles`, `codex`.",
            "schema": {
              "type": "string"
            },
            "example": "profiles,codex"
          }
        ],
        "responses": {
          "200": {
            "description": "DeCC0 entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Decc0Detail"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Your MOCA API key."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your MOCA API key as a bearer token."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid, or revoked API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "errors": [
                {
                  "message": "Invalid or revoked API key.",
                  "extensions": {
                    "code": "UNAUTHORIZED"
                  }
                }
              ]
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded (default 120 requests/minute per key). Check `Retry-After` and the `X-RateLimit-*` headers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Collection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string",
            "description": "Use as the `collection` filter on /v1/artworks"
          },
          "parent_collection": {
            "type": [
              "integer",
              "null"
            ]
          },
          "child_collections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}